refactor: Use types from shared-types instead of openapi
