+++ application/ui/src/features/inference/sinks/local-folder/local-folder.component.tauri.tsx
@@ -6,10 +6,10 @@ import { useCallback, useEffect, useState } from 'react';
 import { Flex, TextField } from '@geti-ui/ui';
 import { open } from '@tauri-apps/plugin-dialog';
 
+import type { LocalFolderSinkConfig } from '../../../../constants/shared-types';
 import { normalizeSelectedPath } from '../../shared/tauri-dialog';
 import { OutputFormats } from '../output-formats/output-formats.component';
 import { RateLimitFields } from '../rate-limit/rate-limit-fields.component';
-import { LocalFolderSinkConfig } from '../utils';
 
 type LocalFolderProps = {
     defaultState: LocalFolderSinkConfig;
