+++ application/ui/src/features/inference/sinks/local-folder/utils.ts
@@ -1,7 +1,8 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { LocalFolderSinkConfig, rateLimitFromFormData, SinkOutputFormats } from '../utils';
+import type { LocalFolderSinkConfig, SinkOutputFormats } from '../../../../constants/shared-types';
+import { rateLimitFromFormData } from '../utils';
 
 export const getLocalFolderInitialConfig = (): LocalFolderSinkConfig => ({
     id: '',
