+++ application/ui/src/features/inference/sinks/mqtt/utils.ts
@@ -1,7 +1,8 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { MqttSinkConfig, rateLimitFromFormData, SinkOutputFormats } from '../utils';
+import type { MqttSinkConfig, SinkOutputFormats } from '../../../../constants/shared-types';
+import { rateLimitFromFormData } from '../utils';
 
 export const getMqttInitialConfig = (): MqttSinkConfig => ({
     id: '',
