+++ application/ui/src/features/inference/sinks/webhook/utils.ts
@@ -1,13 +1,8 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import {
-    getObjectFromFormData,
-    rateLimitFromFormData,
-    SinkOutputFormats,
-    WebhookHttpMethod,
-    WebhookSinkConfig,
-} from '../utils';
+import type { SinkOutputFormats, WebhookSinkConfig } from '../../../../constants/shared-types';
+import { getObjectFromFormData, rateLimitFromFormData, WebhookHttpMethod } from '../utils';
 
 export type Pair = Record<Fields, string>;
