+++ application/ui/src/features/inference/sinks/webhook/webhook.component.tsx
@@ -3,9 +3,10 @@
 
 import { Flex, Item, NumberField, Picker, TextField } from '@geti-ui/ui';
 
+import type { WebhookSinkConfig } from '../../../../constants/shared-types';
 import { OutputFormats } from '../output-formats/output-formats.component';
 import { RateLimitFields } from '../rate-limit/rate-limit-fields.component';
-import { WebhookHttpMethod, WebhookSinkConfig } from '../utils';
+import { WebhookHttpMethod } from '../utils';
 import { HeaderKeyValueBuilder } from './header-key-value-builder.component';
 
 type WebhookProps = {
